refactor(resign): extract resign to param
