+++ test/functions_test.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { strict as assert } from "assert";
-import { passToParam } from "../src/functions.js";
+import { passToParam, resignToParam } from "../src/functions.js";
 
 {
   /** @type {import("../src/functions").Pass} */
@@ -11,4 +11,14 @@ import { passToParam } from "../src/functions.js";
   };
   const expected = "0|p";
   assert.equal(passToParam(input), expected);
+}
+
+{
+  /** @type {import("../src/functions").Resign} */
+  const input = {
+    type: "resign",
+    playerId: 0,
+  };
+  const expected = "0|r";
+  assert.equal(resignToParam(input), expected);
 }
